Reorganized text and code editor files
