+++ Brewmaster/EditorWindows/Code/SourceNavigator.resx
