+++ Brewmaster/EditorWindows/Text/TextEditorWindow.resx
