+++ Brewmaster/ProjectModel/AsmProjectFile.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
-using Brewmaster.EditorWindows;
+using Brewmaster.EditorWindows.Code;
 using Brewmaster.Pipeline;
 
 namespace Brewmaster.ProjectModel
